requirements.txt
forex-python
faker
phonenumbers
reportlab
pyqrcode
pypng
pywhatkit
pillow
pyzbar
selenium
chromedriver-autoinstaller
pyshorteners
speedtest-cli
psutil
wmi
googlesearch-python
python-dotenv